requirements.txt
aiogram==3.22.0
psycopg2-binary==2.9.9
openpyxl==3.1.5
python-dotenv==1.0.1
